+++ connector/query.go
@@ -171,6 +171,12 @@ func prepareElasticsearchQuery(ctx context.Context, request *schema.QueryRequest
 		}
 		if len(aggs) != 0 {
 			query["aggs"] = aggs
+			
+			// set query size to 0 if aggregation is present
+			// this is because, by default, an aggregation query returns both the aggregation result and the hit documents
+			// we only want the aggregation result
+			// more reading: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations.html#return-only-agg-results
+			query["size"] = 0
 		}
 	}