+++ connector/query_test.go
@@ -117,6 +117,18 @@ var tests = []test{
 		group: "customers",
 		name:  "subtype_term_clause",
 	},
+	{
+		group: "payments",
+		name:  "cardinality_aggregation",
+	},
+	{
+		group: "payments",
+		name:  "count_distinct_aggregation",
+	},
+	{
+		group: "payments",
+		name:  "subfield_cardinality_aggregation",
+	},
 }
 
 func TestPrepareElasticsearchQuery(t *testing.T) {